requirements.txt
playwright>=1.40.0
httpx>=0.25.0
trafilatura>=1.6.0
Pillow>=10.0.0
python-dotenv>=1.0.0
lxml-html-clean>=0.1.0
pytest>=8.0.0
pytest-asyncio>=0.23.0